refactor: permission in media-service

+++ media-service/services/attachment/business/delete_attachment.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"strings"
 
+	"thinkflow-service/services/attachment/entity"
+
 	"github.com/VanThen60hz/service-context/core"
 )
 
@@ -12,6 +14,28 @@ func (biz *business) DeleteAttachment(ctx context.Context, id int64) error {
 	attachment, err := biz.attachmentRepo.GetByID(ctx, id)
 	if err != nil {
 		return core.ErrInternalServerError.WithError(err.Error())
+	}
+
+	requester := core.GetRequester(ctx)
+	uid, _ := core.FromBase58(requester.GetSubject())
+	requesterId := int(uid.GetLocalID())
+
+	note, err := biz.noteRepo.GetNoteById(ctx, int(attachment.NoteID))
+	if err != nil {
+		if err == core.ErrRecordNotFound {
+			return core.ErrNotFound.
+				WithError(entity.ErrCannotGetAttachment.Error()).
+				WithDebug(err.Error())
+		}
+
+		return core.ErrInternalServerError.
+			WithError(entity.ErrCannotGetAttachment.Error()).
+			WithDebug(err.Error())
+	}
+
+	if note.UserId != int64(requesterId) {
+		return core.ErrInternalServerError.
+			WithError(entity.ErrRequesterIsNotAttachmentOwner.Error())
 	}
 
 	if err := biz.attachmentRepo.DeleteAttachment(ctx, int(id)); err != nil {
